Add dragging of cards onto the board
Adds slots to the game board which will accept cards. Uses HTML5's
draggable to allow users to drag cards onto the board. Reverts the
Card component to its generic state and creates a new PlayCard
component for use on the board. This component uses individual card
images instead of the spritesheet for simplicity.