Completely refactor the game board
Defines the player and board areas into clean sections without the
jank of absolute positioning. This has the convenient side effect
of fixing the issue of squares not being recognized by react-dnd.

Condenses deck selection and the player's hand into a single component,
and DRYs it up for good measure.

Adds tracking for played cards and properly removes them from the
player's hand when played.