Calculate a winner at the end of the game
Calculates a winner after turn 9 completes. Migrates the setting of
the score state into playCard so that we can check the final values
on the last turn. Adds message timeout as a parameter. Removes the
red/blue turn messages since they're a bit annoying and slow the
game down unnecessarily.