Add Order rule
Requires players to play their cards in deck order from first to last.
Adds a new function to inform Player which cards in the player's
hand are playable based on the active rule. Unplayable cards are
dimmed and cannot be dragged. Cards are also dimmed if it is not
currently the player's turn.